Tutorial creates LBs when requested
The tutorial assumes that services created in K8S are accessible
to the local machine. This may be the case with minikube and nodeport
but is not so with a remotely hosted K8S.

This commit adds functionality to the tutorial that allows a user
to create LB services for the DB and quick start application so that
they are accessible remotely.

The environment variable CREATE_AND_CONFIGURE_LBS must be set to
true to enable this functionality.